Rewrote resolver for accuracy and efficiency
The resolver was rewritten to fix numerous bugs and performance issues.
The reason for this is over the years, bug fixes and features have been
added on top of json-refs and its resolver logic has become crufty.  The
new rewrite is clean, does not reinvent the wheel and fixes many bugs.

Fixes: #80, #87, #88, #89, #97, #100, #101, #103