Issue 125: Fix resolution order
The way json-refs resolved references was top-down.  So as the document
was walked, the reference and its dependencies were recorded.  When
resolving references in this order, if you had a local reference to an
object containing a remote reference, the local reference would be
resolved prior to the remote reference and you could end up with a
partial resolution.

Fixes #125